+++ src/index.ts
@@ -2,3 +2,11 @@
 export * from "./errors/HttpError.js";
 export * from "./errors/CustomError.js";
 export * from "./errors/types.js";
+
+//errors/input-validation
+export * from "./errors/input-validation/InputValidationError.js";
+export * from "./errors/input-validation/zodErrorSerializer.js";
+
+//serializers
+export * from "./serializers/interfaces/IHttpErrorSerializer.js";
+export * from "./serializers/HttpErrorSerializer.js";